==> synthetics/05_compare_associations.ipynb <==
# Compare Associations
* This notebook compares the final GWAS p-values for the full synthetic genome/phenome datasets to those in the original genome/phenome datasets and computes the precision, recall and F1 values.

import os
import pathlib
import pandas as pd

base_path = pathlib.Path(os.getcwd().replace("/synthetics", ""))
data_path = base_path / 'mice_data_set' / 'data' 
real_gwas_path = base_path / 'mice_data_set' / 'out' 
synthetic_gwas_path = base_path / 'mice_data_set' / 'out_synth'


PHENOTYPE = 'abBMD'

real_snps = pd.read_csv(real_gwas_path / f'lm_{PHENOTYPE}_1_79646.csv')
real_snps = real_snps.rename(columns={real_snps.columns[0]: 'index'})
real_snps = real_snps[['index', 'snp', 'p']]
real_snps['interest'] = real_snps['p'].apply(lambda x: True if x <= 1e-8 else False)
real_snps

# Be sure to update the name of your lm file appropriately

synthetic_snps = pd.read_csv(synthetic_gwas_path / 'lm_batchall_abBMD_1_71316.csv') 
synthetic_snps = synthetic_snps.rename(columns={synthetic_snps.columns[0]: 'index'})
synthetic_snps = synthetic_snps[['index', 'snp', 'p']]
synthetic_snps['interest'] = synthetic_snps['p'].apply(lambda x: True if x <= 1e-8 else False)
synthetic_snps

combined = pd.merge(synthetic_snps, 
         real_snps, 
         how='inner', 
         on=['snp'],
         suffixes=['_synthetic', '_real'])
combined

from sklearn.metrics import f1_score, classification_report, confusion_matrix

print(classification_report(combined['interest_real'], combined['interest_synthetic']))
